#26744 Refactor pull operations for separation of concerns and extensibility
Lots of changes have taken place to enhance the modularity and extensibility of the pull methods. Primarily, the pull operations, previously found in FilePullHandler and FilePullController, have now been encapsulated into a separate Puller class for better separation of concerns. This introduces an extra level of abstraction, making the code easier to manage and extend.

The pullTree and pullFile methods from the PullService have also been replaced by a single, more flexible pull method. Additionally, the processing of the file tree, which was initially done in the processTree method of the FilePullHandler, has been moved to the Puller class. This further enhances modularity and improves the overall code structure.

Test files have also been amended to accommodate for these changes. The refactor enhances the code maintainability and allows for easier enhancements and bug fixes in the future.
